+++ changelog.md
@@ -64,6 +64,7 @@
 ## Next
 
 * fileNameAttribute/folderNameAttribute options
+* fileParentAttribute/folderParentAttribute options
 * Tests clear require cache before each test
 * Run jshint on test code
 * Restructure tests